requirements.txt
Django==4.1.3
django_autoslug==1.9.8
